fix: RiskGauge responsive
