+++ vite.config.d.ts
@@ -1,2 +1,2 @@
-declare const _default: any;
+declare const _default: import("vite").UserConfig;
 export default _default;
